+++ src/DataType.php
@@ -31,14 +31,35 @@ enum DataType: string
     /**
      * @throws StructuredFieldError
      */
-    public function build(iterable $data): string
+    public function serialize(iterable $data): string
     {
-        return (match ($this) {
+        return $this->create($data)->toHttpValue();
+    }
+
+    /**
+     * @throws StructuredFieldError
+     */
+    public function create(iterable $data): StructuredField
+    {
+        return match ($this) {
             self::List => OuterList::fromPairs($data),
             self::InnerList => InnerList::fromPair([...$data]), /* @phpstan-ignore-line */
             self::Parameters => Parameters::fromPairs($data),
             self::Dictionary => Dictionary::fromPairs($data),
             self::Item => Item::fromPair([...$data]), /* @phpstan-ignore-line */
-        })->toHttpValue();
+        };
+    }
+
+    /**
+     * DEPRECATION WARNING! This method will be removed in the next major point release.
+     *
+     * @deprecated Since version 1.3.0
+     * @codeCoverageIgnore
+     *
+     * @see DataType::serialize()
+     */
+    public function build(iterable $data): string
+    {
+        return $this->serialize($data);
     }
 }